# inspection/upper_computer/frontend/openapi/inspection_gateway_openapi.json
{
  "components": {
    "schemas": {
...
      "EmptyActionRequest": {
        "additionalProperties": false,
        "description": "Empty payload used by actions that carry no caller-provided parameters.",
        "properties": {},
        "title": "EmptyActionRequest",
        "type": "object"
      },
      "ExecuteReplayRequest": {
        "additionalProperties": false,
        "description": "Replay one persisted inspection trace.",
        "properties": {
          "traceId": {
            "description": "Persisted trace identifier to replay.",
            "minLength": 1,
            "title": "Traceid",
            "type": "string"
          }
        },
        "required": [
          "traceId"
        ],
        "title": "ExecuteReplayRequest",
        "type": "object"
      },
      "ExportBatchRequest": {
        "additionalProperties": false,
        "description": "Export one persisted batch bundle.",
        "properties": {
          "batchId": {
            "description": "Batch identifier to export.",
            "minLength": 1,
            "title": "Batchid",
            "type": "string"
          }
        },
        "required": [
          "batchId"
        ],
        "title": "ExportBatchRequest",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MaintenanceModeRequest": {
        "additionalProperties": false,
        "description": "Switch maintenance mode through the canonical action plane.",
        "properties": {
          "enabled": {
            "description": "Target maintenance-mode state.",
            "title": "Enabled",
            "type": "boolean"
          }
        },
        "required": [
          "enabled"
        ],
        "title": "MaintenanceModeRequest",
        "type": "object"
      },
...
      "ResetStationRequest": {
        "additionalProperties": false,
        "description": "Reset one station fault and optionally resume automation.",
        "properties": {
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional operator reason stored with the reset request.",
            "title": "Reason"
          },
          "resumeAfter": {
            "default": false,
            "description": "Whether the station should resume automation after the fault reset completes.",
            "title": "Resumeafter",
            "type": "boolean"
          }
        },
        "title": "ResetStationRequest",
        "type": "object"
      },
...
      "StartBatchRequest": {
        "additionalProperties": false,
        "description": "Start one production batch.\n\nArgs:\n    recipeId: Activated recipe identifier to execute.\n    batchId: Optional caller-provided batch identifier.",
        "properties": {
          "batchId": {
            "anyOf": [
              {
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional caller-provided batch identifier.",
            "title": "Batchid"
          },
          "recipeId": {
            "description": "Activated recipe identifier used by the next batch run.",
            "minLength": 1,
            "title": "Recipeid",
            "type": "string"
          }
        },
        "required": [
          "recipeId"
        ],
        "title": "StartBatchRequest",
        "type": "object"
      },
      "SwitchRecipeRequest": {
        "additionalProperties": false,
        "description": "Validate one recipe and optionally activate it for the next run.",
        "properties": {
          "dryRun": {
            "default": false,
            "description": "When true, validate the recipe without activating it.",
            "title": "Dryrun",
            "type": "boolean"
          },
          "recipeId": {
            "description": "Recipe identifier to validate and optionally activate.",
            "minLength": 1,
            "title": "Recipeid",
            "type": "string"
          }
        },
        "required": [
          "recipeId"
        ],
        "title": "SwitchRecipeRequest",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Inspection HMI Gateway",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/actions/capability-matrix": {
      "get": {
        "operationId": "getActionCapabilityMatrix",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Getactioncapabilitymatrix",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get the effective action capability matrix",
        "tags": [
          "actions"
        ]
      }
    },
    "/api/v1/actions/catalog": {
      "get": {
        "operationId": "getActionCatalog",
        "parameters": [
          {
            "in": "query",
            "name": "include_non_production",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Non Production",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Getactioncatalog",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List the public action catalog",
        "tags": [
          "actions"
        ]
      }
    },
    "/api/v1/actions/create-batch": {
      "post": {
        "operationId": "submitCreateBatchAction",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmptyActionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Submitcreatebatchaction",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Submit the create-batch action",
        "tags": [
          "actions"
        ]
      }
    },
    "/api/v1/actions/diagnostics/capture-frame": {
      "post": {
        "operationId": "submitDiagnosticCaptureFrameAction",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmptyActionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Submitdiagnosticcaptureframeaction",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Submit the capture-frame diagnostic action",
        "tags": [
          "actions"
        ]
      }
    },
    "/api/v1/actions/diagnostics/test-lighting": {
      "post": {
        "operationId": "submitDiagnosticTestLightingAction",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmptyActionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Submitdiagnostictestlightingaction",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Submit the lighting diagnostic action",
        "tags": [
          "actions"
        ]
      }
    },
    "/api/v1/actions/diagnostics/test-sort-actuator": {
      "post": {
        "operationId": "submitDiagnosticTestSortActuatorAction",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmptyActionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Submitdiagnostictestsortactuatoraction",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Submit the sort-actuator diagnostic action",
        "tags": [
          "actions"
        ]
      }
    },
    "/api/v1/actions/execute-replay": {
      "post": {
        "operationId": "submitExecuteReplayAction",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExecuteReplayRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Submitexecutereplayaction",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Submit the trace replay action",
        "tags": [
          "actions"
        ]
      }
    },
    "/api/v1/actions/export-batch": {
      "post": {
        "operationId": "submitExportBatchAction",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExportBatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Submitexportbatchaction",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Submit the batch export action",
        "tags": [
          "actions"
        ]
      }
    },
    "/api/v1/actions/jobs": {
      "get": {
        "operationId": "listActionJobs",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 1000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Listactionjobs",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List action jobs",
        "tags": [
          "actions"
        ]
      }
    },
    "/api/v1/actions/jobs/{job_id}": {
      "get": {
        "operationId": "getActionJob",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Getactionjob",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get one action job",
        "tags": [
          "actions"
        ]
      }
    },
    "/api/v1/actions/jobs/{job_id}/cancel": {
      "post": {
        "operationId": "cancelActionJob",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Cancelactionjob",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Cancel one action job",
        "tags": [
          "actions"
        ]
      }
    },
    "/api/v1/actions/reset-station": {
      "post": {
        "operationId": "submitResetStationAction",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetStationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Submitresetstationaction",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Submit the reset-station action",
        "tags": [
          "actions"
        ]
      }
    },
    "/api/v1/actions/set-maintenance-mode": {
      "post": {
        "operationId": "submitMaintenanceModeAction",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MaintenanceModeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Submitmaintenancemodeaction",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Submit the maintenance-mode action",
        "tags": [
          "actions"
        ]
      }
    },
    "/api/v1/actions/start-batch": {
      "post": {
        "operationId": "submitStartBatchAction",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartBatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Submitstartbatchaction",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Submit the start-batch action",
        "tags": [
          "actions"
        ]
      }
    },
    "/api/v1/actions/stop-station": {
      "post": {
        "operationId": "submitStopStationAction",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmptyActionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Submitstopstationaction",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Submit the stop-station action",
        "tags": [
          "actions"
        ]
      }
    },
    "/api/v1/actions/switch-recipe": {
      "post": {
        "operationId": "submitSwitchRecipeAction",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SwitchRecipeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Submitswitchrecipeaction",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Submit the recipe validation and activation action",
        "tags": [
          "actions"
        ]
      }
    },
...
  }
}